New indicator sample: Sample1Blank
