--- project/OsEngine/Indicators/Samples/Sample2IndicatorDataSeries.cs
@@ -1,27 +1,0 @@
-﻿using OsEngine.Entity;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace OsEngine.Indicators.Samples
-{
-    [Indicator("Sample2IndicatorDataSeries")]
-    public class Sample2IndicatorDataSeries : Aindicator
-    {
-        public override void OnStateChange(IndicatorState state)
-        {
-            
-
-
-        }
-
-        public override void OnProcess(List<Candle> source, int index)
-        {
-            
-
-
-        }
-    }
-}
