+++ project/OsEngine/Indicators/Samples/Sample3IndicatorDataSeries.cs
@@ -1,0 +1,27 @@
+﻿using OsEngine.Entity;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace OsEngine.Indicators.Samples
+{
+    [Indicator("Sample3IndicatorDataSeries")]
+    public class Sample3IndicatorDataSeries : Aindicator
+    {
+        public override void OnStateChange(IndicatorState state)
+        {
+            
+
+
+        }
+
+        public override void OnProcess(List<Candle> source, int index)
+        {
+            
+
+
+        }
+    }
+}
